add:FrameBeachを追加
